+++ packages/browser-wallet-api-helpers/CHANGELOG.md
@@ -3,6 +3,7 @@
 ### Unreleased
 
 -   Updated handler `sendTransaction` with new payload type `TokenUpdatePayload`
+-   Added `requestVerifiablePresentationV1` entrypoint to prove statements about identities and verifiable credentials with auditable anchor
 
 ## 3.0.1