+++ packages/browser-wallet-api-helpers/CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ### Unreleased
 
+### 3.1.0
+
 -   Updated handler `sendTransaction` with new payload type `TokenUpdatePayload`
 
 ## 3.0.1